+++ packages/client/src/components/modal/modal.scss
@@ -1,51 +1,3 @@
-// .modal{
-//   position: fixed;
-//   height: 100vh;
-//   width: 100vw;
-//   top: 0;
-//   left: 0;
-//   display: flex;
-//   align-items: center;
-//   justify-content: center;
-//   z-index: 2;
-// }
-
-// .modal-mask{
-//     position: absolute;
-//     width: 100vw;
-//     height: 100vh;
-//     background: #212B799C;
-//   }
-  
-//   .modal-view{
-//     background: #CCD6FE;
-//     z-index: 2;
-//     width: 100%;
-//     max-width: 420px;
-//     min-height: 300px;
-//     height: auto;
-//     border-radius: 12px;
-//     position: relative;
-//     display: flex;
-//     flex-direction: column;
-//     align-items: center;
-//     justify-content: center;
-//     padding: 41px 50px 37px 50px
-//   }
-  
-//   .modal-view-close{
-//     position: absolute;
-//     top:15px;
-//     right: 15px;
-//     height: 14px;
-//     width: 14px;
-//     cursor: pointer;
-//   }
-
-//   .mb-28{
-//     margin-bottom: 28px;
-//   }
-
 @import "./../../scss/base/variables";
 
 .modal {
@@ -128,4 +80,8 @@
   flex-direction: column;
   align-items: stretch;
   gap: 28px;
+}
+
+.mb-28{
+  margin-bottom: 28px;
 }
